% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPAS.R
\name{findPAS}
\alias{findPAS}
\title{PAS peaks identification}
\usage{
findPAS(
  samplenames,
  chrs,
  outpath,
  threads,
  cache.covdata = F,
  callpeak = "macs3",
  genomeSize = 2.7e+09,
  cutOff = 10^-5,
  additionalParams = "--nomodel",
  min.gapwidth = 50,
  sum.count = 10,
  min.count = 10,
  cpm.cutoff = 0.5,
  BSg = BSgenome.Hsapiens.UCSC.hg38,
  literal = "AAAAAAAA",
  num_A = 10
)
}
\arguments{
\item{samplenames}{A vector containing sample names.}

\item{chrs}{A vector containing chromatin to use.}

\item{outpath}{A character describing where the file will be saved.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{cache.covdata}{A boolean value that determines whether cache coverage file shoulde be used.}

\item{callpeak}{Method to identify PAS peak. Options include "macs2","macs3", or "trim".}

\item{genomeSize}{The number of genome size.}

\item{cutOff}{The number of peak threshold for macs-based mathod.}

\item{additionalParams}{Additional params for macs-based mathod.}

\item{min.gapwidth}{The minimum distance between PAS peaks. Peaks gaps smaller than this distance will be merged.}

\item{sum.count}{The minimum count of PAS peak. Peaks count smaller than this will be removed.}

\item{min.count}{The minimum count of PAS peak identification for "trim" method.}

\item{cpm.cutoff}{The minimum CPM normalized count of PAS peak. PAS peaks with a count smaller than this will be removed.}

\item{literal}{If the literal sequence is detected within 5bp downstream and 20bp upstream of PAS 3' end, the PAS peak will be deleted.}

\item{num_A}{If more than the number of A base is detected within 5bp downstream and 20bp upstream of PAS 3' end, the PAS peak will be deleted.}
}
\description{
PAS peaks identification
}
