% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeAnno.R
\name{MakeAnnoFromGtf}
\alias{MakeAnnoFromGtf}
\title{Make gene annotation GRanges From gtf}
\usage{
MakeAnnoFromGtf(gtf)
}
\arguments{
\item{gtf}{Input gtf file}
}
\description{
Make gene annotation GRanges From gtf
}
