% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAisoformCount.R
\name{pas_fit_model}
\alias{pas_fit_model}
\title{Model fitting}
\usage{
pas_fit_model(
  samplenames,
  chrs,
  outpath,
  gr_pas,
  cache.dist = FALSE,
  intronanno,
  threads = 24,
  intron.remove = TRUE,
  sampling.frac = 4/100
)
}
\arguments{
\item{samplenames}{A vector containing sample names.}

\item{chrs}{A vector containing chromatin to use.}

\item{outpath}{A character describing where the file will be saved.}

\item{gr_pas}{A range of PAS peak.}

\item{cache.dist}{A boolean value that determines whether cache distance file shoulde be used.}

\item{intronanno}{A ranges describing intron annotation.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{intron.remove}{A boolean value that determines whether remove intron when caculate the distance between Read2 and PAS.}

\item{sampling.frac}{The proportion of reads used for distance fitting.}
}
\description{
Model fitting
}
