% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPAS.R
\name{pas_peak2gene}
\alias{pas_peak2gene}
\title{Link PAS peaks to genes}
\usage{
pas_peak2gene(gr_pas, samplenames, chrs, outpath, threads = 24)
}
\arguments{
\item{gr_pas}{GRanges of PAS peaks.}

\item{samplenames}{Vector of sample names processed by `ProcessBam`.}

\item{chrs}{Chromosomes included in the analysis.}

\item{outpath}{Directory where bam2df/pas.df intermediates reside.}

\item{threads}{Number of threads for processing.}
}
\description{
Link PAS peaks to genes
}
