% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPAS.R
\name{pas_peak_anno}
\alias{pas_peak_anno}
\title{pas peak annotation}
\usage{
pas_peak_anno(
  gr_pas = pas_peak,
  gene.anno,
  priority = c("utr3", "last.exon", "last.intron", "exon", "intron", "utr5"),
  outpath
)
}
\arguments{
\item{gr_pas}{A range of PAS peak.}

\item{gene.anno}{A ranges list describing 3' UTR, exon, and intron.}

\item{priority}{Annotation priority order.}

\item{outpath}{Output directory where peakanno.pdf will be written.}
}
\description{
pas peak annotation
}
