% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAisoformCount.R
\name{pas_predict}
\alias{pas_predict}
\title{PA isoform quantification}
\usage{
pas_predict(
  gr_pas,
  samplenames,
  chrs,
  outpath,
  geneanno,
  intronanno,
  intron.remove = TRUE,
  remove.bad = "dist",
  cell.metadata,
  fitmodel = "gamma",
  threads = 24
)
}
\arguments{
\item{gr_pas}{A range of PAS peak.}

\item{samplenames}{A vector containing sample names.}

\item{chrs}{A vector containing chromatin to use.}

\item{outpath}{A character describing where the file will be saved.}

\item{geneanno}{A ranges describing gene annotation.}

\item{intronanno}{A ranges describing intron annotation.}

\item{intron.remove}{A boolean value that determines whether remove intron when caculate the distance between Read2 and PAS.}

\item{cell.metadata}{A data.frame containing: cb, cell barcode id; celltype, cell type annotation; samplename, sample name.}

\item{fitmodel}{The distribution used for distance fitting. Options include "weibull", "gamma", or "lognormal".}

\item{threads}{The number of threads to be used for parallel computing.}
}
\description{
PA isoform quantification
}
