% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAisoformCount.R
\name{PasUsageScore}
\alias{PasUsageScore}
\title{Model fitting}
\usage{
PasUsageScore(
  sc.pas = pas_sc_count,
  gr_pas,
  min.pas.cutoff = 5,
  pseudo_count = 0.1,
  threads = 24,
  by = "utr",
  out.matrix = F
)
}
\arguments{
\item{sc.pas}{PAS count matrix: could be PAS peak X cell matrix or PAS peak X celltype.}

\item{gr_pas}{A range of PAS peak.}

\item{min.pas.cutoff}{If the rowSums(PAS count) of a PAS peak is less than this value, it will not be used for calculation.}

\item{pseudo_count}{The number of count that will be add to PAS count.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{out.matrix}{A boolean value that determines whether  to export the APA rank score of each 3' UTR region in each cell or celltype.}
}
\description{
Model fitting
}
