% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vis.R
\name{plot_base_freq}
\alias{plot_base_freq}
\title{Plot nucleotide composition around PAS peaks}
\usage{
plot_base_freq(
  gr_pas,
  extend_base = 100,
  BSg = default_hg38_bsgenome(),
  outpath,
  outname = "pas.100bp.freq.pdf",
  ylim = 0.8
)
}
\arguments{
\item{gr_pas}{GRanges of PAS peaks that include the `WhichMaxs` metadata column.}

\item{extend_base}{Number of bases to extend upstream/downstream of each peak apex.}

\item{BSg}{BSgenome reference used to extract sequences.}

\item{outpath}{Directory where the PDF will be written.}

\item{outname}{Output filename (relative to `outpath`).}

\item{ylim}{Upper bound for the fraction axis.}
}
\description{
Plot nucleotide composition around PAS peaks
}
