% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\docType{package}
\name{scPAISO}
\alias{scPAISO-package}
\alias{scPAISO}
\title{scPAISO package}
\description{
Utilities for identifying and quantifying alternative polyadenylation sites
using single-cell RNA-seq data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yongjieliu/scPAISO}
  \item Report bugs at \url{https://github.com/yongjieliu/scPAISO/issues}
}

}
\keyword{internal}
