% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPAS.R
\name{pas_peak2gene}
\alias{pas_peak2gene}
\title{link pas peak to gene}
\usage{
pas_peak2gene(
  gr_pas = pas_peak,
  samplenames = samplenames,
  chrs = chrs,
  threads = 24
)
}
\arguments{
\item{gr_pas}{A range of PAS peak.}

\item{samplenames}{A vector containing sample names.}

\item{chrs}{A vector containing chromatin to use.}

\item{threads}{The number of threads to be used for parallel computing.}
}
\description{
link pas peak to gene
}
