% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPAS.R
\name{pas_peak_anno}
\alias{pas_peak_anno}
\title{Annotate PAS peaks}
\usage{
pas_peak_anno(
  gr_pas,
  gene.anno,
  priority = c("utr3", "last.exon", "last.intron", "exon", "intron", "utr5"),
  outpath
)
}
\arguments{
\item{gr_pas}{GRanges of PAS peaks annotated with gene IDs.}

\item{gene.anno}{List returned by `MakeAnnoFromGtf`.}

\item{priority}{Priority order of genomic region labels.}

\item{outpath}{Directory where `peakanno.pdf` should be saved.}
}
\description{
Annotate PAS peaks
}
