% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPAS.R
\name{pas_peak_anno}
\alias{pas_peak_anno}
\title{pas peak annotation}
\usage{
pas_peak_anno(
  gr_pas = pas_peak,
  gene.anno,
  priority = c("utr3", "last.exon", "last.intron", "exon", "intron", "utr5")
)
}
\arguments{
\item{gr_pas}{A range of PAS peak.}

\item{gene.anno}{A ranges list describing 3' UTR, exon, and intron.}

\item{chrs}{A vector containing chromatin to use.}

\item{threads}{The number of threads to be used for parallel computing.}
}
\description{
pas peak annotation
}
